abstract channel stub
