+++ tests/channel.js
@@ -1,0 +1,1 @@
+import { AbstractChannel } from '../dist/index.js';
